Module aws_sdk_databrew::model

Modules

column_selector

See ColumnSelector

column_statistics_configuration

See ColumnStatisticsConfiguration

condition_expression

See ConditionExpression

csv_options

See CsvOptions

csv_output_options

See CsvOutputOptions

data_catalog_input_definition

See DataCatalogInputDefinition

data_catalog_output

See DataCatalogOutput

database_input_definition

See DatabaseInputDefinition

database_output

See DatabaseOutput

database_table_output_options

See DatabaseTableOutputOptions

dataset

See Dataset

dataset_parameter

See DatasetParameter

datetime_options

See DatetimeOptions

excel_options

See ExcelOptions

files_limit

See FilesLimit

filter_expression

See FilterExpression

format_options

See FormatOptions

input

See Input

job

See Job

job_run

See JobRun

job_sample

See JobSample

json_options

See JsonOptions

output

See Output

output_format_options

See OutputFormatOptions

path_options

See PathOptions

profile_configuration

See ProfileConfiguration

project

See Project

recipe

See Recipe

recipe_action

See RecipeAction

recipe_reference

See RecipeReference

recipe_step

See RecipeStep

recipe_version_error_detail

See RecipeVersionErrorDetail

s3_location

See S3Location

s3_table_output_options

See S3TableOutputOptions

sample

See Sample

schedule

See Schedule

statistic_override

See StatisticOverride

statistics_configuration

See StatisticsConfiguration

view_frame

See ViewFrame

Structs

ColumnSelector

Selector of a column from a dataset for profile job configuration. One selector includes either a column name or a regular

ColumnStatisticsConfiguration

Configuration for column evaluations for a profile job. ColumnStatisticsConfiguration can be used to select

ConditionExpression

Represents an individual condition that evaluates to true or false.

CsvOptions

Represents a set of options that define how DataBrew will read a

CsvOutputOptions

Represents a set of options that define how DataBrew will write a

DataCatalogInputDefinition

Represents how metadata stored in the Glue Data Catalog is defined in a DataBrew

DataCatalogOutput

Represents options that specify how and where in the Glue Data Catalog DataBrew

DatabaseInputDefinition

Connection information for dataset input files stored in a database.

DatabaseOutput

Represents a JDBC database output object which defines the output destination for

DatabaseTableOutputOptions

Represents options that specify how and where DataBrew writes the database output generated by recipe

Dataset

Represents a dataset that can be processed by DataBrew.

DatasetParameter

Represents a dataset paramater that defines type and conditions for a parameter in the Amazon S3

DatetimeOptions

Represents additional options for correct interpretation of datetime parameters used in

ExcelOptions

Represents a set of options that define how DataBrew will interpret a Microsoft Excel file when

FilesLimit

Represents a limit imposed on number of Amazon S3 files that should be selected for a dataset from a connected

FilterExpression

Represents a structure for defining parameter conditions. Supported conditions are described

FormatOptions

Represents a set of options that define the structure of either comma-separated value (CSV),

Input

Represents information on how DataBrew can find data, in either the Glue Data Catalog or

Job

Represents all of the attributes of a DataBrew job.

JobRun

Represents one run of a DataBrew job.

JobSample

A sample configuration for profile jobs only, which determines the number of rows on which the

JsonOptions

Represents the JSON-specific options that define how input is to be interpreted by Glue DataBrew.

Output

Represents options that specify how and where in Amazon S3 DataBrew writes the output generated by

OutputFormatOptions

Represents a set of options that define the structure of comma-separated (CSV) job output.

PathOptions

Represents a set of options that define how DataBrew selects files for a given Amazon S3

ProfileConfiguration

Configuration for profile jobs. Configuration can be used to select columns, do evaluations, and override default

Project

Represents all of the attributes of a DataBrew project.

Recipe

Represents one or more actions to be performed on a DataBrew dataset.

RecipeAction

Represents a transformation and associated parameters that are used to apply a change

RecipeReference

Represents the name and version of a DataBrew recipe.

RecipeStep

Represents a single step from a DataBrew recipe to be performed.

RecipeVersionErrorDetail

Represents any errors encountered when attempting to delete multiple recipe

S3Location

Represents an Amazon S3 location (bucket name and object key) where DataBrew can read

S3TableOutputOptions

Represents options that specify how and where DataBrew writes the Amazon S3 output generated by

Sample

Represents the sample size and sampling type for DataBrew to use for interactive data

Schedule

Represents one or more dates and times when a job is to run.

StatisticOverride

Override of a particular evaluation for a profile job.

StatisticsConfiguration

Configuration of evaluations for a profile job. This configuration can be used to select

ViewFrame

Represents the data being transformed during an action.

Enums

CompressionFormat

DatabaseOutputMode

EncryptionMode

InputFormat

JobRunState

JobType

LogSubscription

Order

OrderedBy

OutputFormat

ParameterType

SampleMode

SampleType

SessionStatus

Source