Module aws_sdk_databrew::operation

Structs

BatchDeleteRecipeVersion

Deletes one or more versions of a recipe at a time.

CreateDataset

Creates a new DataBrew dataset.

CreateProfileJob

Creates a new job to analyze a dataset and create its data profile.

CreateProject

Creates a new DataBrew project.

CreateRecipe

Creates a new DataBrew recipe.

CreateRecipeJob

Creates a new job to transform input data, using steps defined in an existing Glue DataBrew recipe

CreateSchedule

Creates a new schedule for one or more DataBrew jobs. Jobs can be run at a specific

DeleteDataset

Deletes a dataset from DataBrew.

DeleteJob

Deletes the specified DataBrew job.

DeleteProject

Deletes an existing DataBrew project.

DeleteRecipeVersion

Deletes a single version of a DataBrew recipe.

DeleteSchedule

Deletes the specified DataBrew schedule.

DescribeDataset

Returns the definition of a specific DataBrew dataset.

DescribeJob

Returns the definition of a specific DataBrew job.

DescribeJobRun

Represents one run of a DataBrew job.

DescribeProject

Returns the definition of a specific DataBrew project.

DescribeRecipe

Returns the definition of a specific DataBrew recipe corresponding to a particular

DescribeSchedule

Returns the definition of a specific DataBrew schedule.

ListDatasets

Lists all of the DataBrew datasets.

ListJobRuns

Lists all of the previous runs of a particular DataBrew job.

ListJobs

Lists all of the DataBrew jobs that are defined.

ListProjects

Lists all of the DataBrew projects that are defined.

ListRecipeVersions

Lists the versions of a particular DataBrew recipe, except for

ListRecipes

Lists all of the DataBrew recipes that are defined.

ListSchedules

Lists the DataBrew schedules that are defined.

ListTagsForResource

Lists all the tags for a DataBrew resource.

PublishRecipe

Publishes a new version of a DataBrew recipe.

SendProjectSessionAction

Performs a recipe step within an interactive DataBrew session that's currently

StartJobRun

Runs a DataBrew job.

StartProjectSession

Creates an interactive session, enabling you to manipulate data in a DataBrew

StopJobRun

Stops a particular run of a job.

TagResource

Adds metadata tags to a DataBrew resource, such as a dataset, project, recipe, job, or

UntagResource

Removes metadata tags from a DataBrew resource.

UpdateDataset

Modifies the definition of an existing DataBrew dataset.

UpdateProfileJob

Modifies the definition of an existing profile job.

UpdateProject

Modifies the definition of an existing DataBrew project.

UpdateRecipe

Modifies the definition of the LATEST_WORKING version of a DataBrew

UpdateRecipeJob

Modifies the definition of an existing DataBrew recipe job.

UpdateSchedule

Modifies the definition of an existing DataBrew schedule.